App: Move Filter classes to '.settings.model' package
