--- app/src/main/java/soup/movie/data/model/GenreFilter.kt
@@ -1,5 +1,0 @@
-package soup.movie.data.model
-
-data class GenreFilter(
-    val blacklist: Set<String>
-)
