+++ app/src/main/java/soup/movie/settings/impl/TheaterFilterSetting.kt
@@ -1,7 +1,7 @@
 package soup.movie.settings.impl
 
 import android.content.SharedPreferences
-import soup.movie.data.model.TheaterFilter
+import soup.movie.settings.model.TheaterFilter
 import soup.movie.settings.PrefSetting
 
 class TheaterFilterSetting(
