+++ app/src/main/java/soup/movie/settings/model/AgeFilter.kt
@@ -1,6 +1,6 @@
-package soup.movie.data.model
+package soup.movie.settings.model
 
-data class AgeFilter(private val flags: Int) {
+class AgeFilter(private val flags: Int) {
 
     fun hasAll(): Boolean = has(FLAG_AGE_ALL)
     fun has12(): Boolean = has(FLAG_AGE_12)
@@ -19,7 +19,6 @@ data class AgeFilter(private val flags: Int) {
         const val FLAG_AGE_12: Int = 0x2
         const val FLAG_AGE_15: Int = 0x4
         const val FLAG_AGE_19: Int = 0x8
-        const val FLAG_AGE_DEFAULT: Int =
-                FLAG_AGE_ALL or FLAG_AGE_12 or FLAG_AGE_15 or FLAG_AGE_19
+        const val FLAG_AGE_DEFAULT: Int = FLAG_AGE_ALL or FLAG_AGE_12 or FLAG_AGE_15 or FLAG_AGE_19
     }
 }
