+++ app/src/main/java/soup/movie/settings/model/GenreFilter.kt
@@ -1,0 +1,5 @@
+package soup.movie.settings.model
+
+class GenreFilter(
+    val blacklist: Set<String>
+)
