+++ app/src/main/java/soup/movie/ui/home/filter/HomeFilterViewModel.kt
@@ -5,16 +5,16 @@ import androidx.lifecycle.MutableLiveData
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.rxkotlin.Observables
 import io.reactivex.schedulers.Schedulers
-import soup.movie.data.model.AgeFilter
-import soup.movie.data.model.AgeFilter.Companion.FLAG_AGE_12
-import soup.movie.data.model.AgeFilter.Companion.FLAG_AGE_15
-import soup.movie.data.model.AgeFilter.Companion.FLAG_AGE_19
-import soup.movie.data.model.AgeFilter.Companion.FLAG_AGE_ALL
-import soup.movie.data.model.GenreFilter
-import soup.movie.data.model.TheaterFilter
-import soup.movie.data.model.TheaterFilter.Companion.FLAG_THEATER_CGV
-import soup.movie.data.model.TheaterFilter.Companion.FLAG_THEATER_LOTTE
-import soup.movie.data.model.TheaterFilter.Companion.FLAG_THEATER_MEGABOX
+import soup.movie.settings.model.AgeFilter
+import soup.movie.settings.model.AgeFilter.Companion.FLAG_AGE_12
+import soup.movie.settings.model.AgeFilter.Companion.FLAG_AGE_15
+import soup.movie.settings.model.AgeFilter.Companion.FLAG_AGE_19
+import soup.movie.settings.model.AgeFilter.Companion.FLAG_AGE_ALL
+import soup.movie.settings.model.GenreFilter
+import soup.movie.settings.model.TheaterFilter
+import soup.movie.settings.model.TheaterFilter.Companion.FLAG_THEATER_CGV
+import soup.movie.settings.model.TheaterFilter.Companion.FLAG_THEATER_LOTTE
+import soup.movie.settings.model.TheaterFilter.Companion.FLAG_THEATER_MEGABOX
 import soup.movie.domain.filter.GetGenreListUseCase
 import soup.movie.settings.impl.AgeFilterSetting
 import soup.movie.settings.impl.GenreFilterSetting
